% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np.mojo.R
\name{np.mojo}
\alias{np.mojo}
\title{Nonparametric Single Lag Change Point Detection for Multivariate Time Series}
\usage{
np.mojo(
  x,
  G,
  lag = 0,
  kernel.f = c("quad.exp", "gauss", "euclidean", "laplace", "sine")[1],
  kern.par = 1,
  data.driven.kern.par = TRUE,
  alpha = 0.1,
  threshold = c("bootstrap", "manual")[1],
  threshold.val = NULL,
  reps = 199,
  boot.dep = 1.5 * (nrow(as.matrix(x))^(1/3)),
  parallel = FALSE,
  boot.method = c("mean.subtract", "no.mean.subtract")[1],
  criterion = c("eta", "epsilon", "eta.and.epsilon")[3],
  eta = 0.4,
  epsilon = 0.02,
  use.mean = FALSE
)
}
\arguments{
\item{x}{Input data (a \code{numeric} vector or an object of classes \code{ts} and \code{timeSeries},
or a \code{numeric} matrix with rows representing variables).}

\item{G}{An integer value for the moving sum bandwidth;
\code{G} should be less than \code{length(n)/2}.}

\item{lag}{The lagged values of the time series used to detect changes. If \code{lag} \eqn{\ell = 0}, then only marginal changes are detected.
If \code{lag} \eqn{\ell \neq 0}, then changes in the pairwise distribution of \eqn{(X_t , X_{t+\ell})} are detected.}

\item{kernel.f}{String indicating which kernel function to use when calculating the NP-MOJO detectors statistics; with \code{kern.par} \eqn{= a}, possible values are
 \itemize{
   \item{\code{"quad.exp"}}{: kernel \eqn{h_2} in McGonigle and Cho (2023), kernel 5 in Fan et al. (2017):
   \deqn{h (x,y) = \prod_{i=1}^{2p} \frac{ (2a - (x_i - y_i)^2) \exp (-\frac{1}{4a} (x_i - y_i)^2 )}{2a} .}  }
   \item{\code{"gauss"}}{: kernel \eqn{h_1} in McGonigle and Cho (2023), the standard Gaussian kernel:
   \deqn{h (x,y) = \exp ( - \frac{a^2}{2} \Vert x - y  \Vert^2) .} }
   \item{\code{"euclidean"}}{: kernel \eqn{h_3} in McGonigle and Cho (2023), the Euclidean distance-based kernel:
   \deqn{h (x, y ) = \Vert x - y \Vert^a  .}}
   \item{\code{"laplace"}}{: kernel 2 in Fan et al. (2017), based on a Laplace weight function:
     \deqn{h (x, y ) = \prod_{i=1}^{2p} \left( 1+ a^2 (x_i - y_i)^2  \right)^{-1}. }}
   \item{\code{"sine"}}{: kernel 4 in Fan et al. (2017), based on a sinusoidal weight function:
     \deqn{h (x, y ) = \prod_{i=1}^{2p} \frac{-2 | x_i - y_i |  + | x_i - y_i - 2a|  + | x_i - y_i +2a| }{4a} .}}
}}

\item{kern.par}{The tuning parameter that appears in the expression for the kernel function, which acts as a scaling parameter,
only to be used if \code{data.driven.kern.par = FALSE}. If \code{kernel.f = "euclidean"}, then \code{kern.par} \eqn{\in (0,2)},
otherwise \code{kern.par} \eqn{> 0}.}

\item{data.driven.kern.par}{A \code{logical} variable, if set to \code{TRUE}, then the kernel tuning parameter is calculated
using the median heuristic, if \code{FALSE} it is given by \code{kern.par}.}

\item{alpha}{A numeric value for the significance level with
\code{0 <= alpha <= 1}; use iff \code{threshold = "bootstrap"}.}

\item{threshold}{String indicating how the threshold is computed. Possible values are
 \itemize{
   \item{\code{"bootstrap"}}{: the threshold is calculated using the bootstrap method with significance level \code{alpha}.}
       \item{\code{"manual"}}{: the threshold is set by the user and must be specified using the \code{threshold.val} parameter.}
}}

\item{threshold.val}{The value of the threshold used to declare change points, only to be used if \code{threshold = "manual"}.}

\item{reps}{An integer value for the number of bootstrap replications performed, if \code{threshold = "bootstrap"}.}

\item{boot.dep}{A positive value for the strength of dependence in the multiplier bootstrap sequence, if \code{threshold = "bootstrap"}.}

\item{parallel}{A \code{logical} variable, if set to \code{TRUE}, then parallel computing is used in the bootstrapping procedure
if bootstrapping is performed.}

\item{boot.method}{A string indicating the method for creating bootstrap replications. It is not recommended to change this. Possible choices are
\itemize{
   \item{\code{"mean.subtract"}}{: the default choice, as described in McGonigle and Cho (2023). Empirical mean subtraction is performed to the bootstrapped replicates, improving power.}
       \item{\code{"no.mean.subtract"}}{: empirical mean subtraction is not performed, improving size control.}
}}

\item{criterion}{String indicating how to determine whether each point \code{k} at which NP-MOJO statistic
exceeds the threshold is a change point; possible values are
 \itemize{
   \item{\code{"epsilon"}}{: \code{k} is the maximum of its local exceeding environment, which has at least size \code{epsilon*G}.}
       \item{\code{"eta"}}{: there is no larger exceeding in an \code{eta*G} environment of \code{k}.}
       \item{\code{"eta.and.epsilon"}}{: the recommended default option; \code{k} satisfies both the eta and epsilon criterion.
       Recommended to use with the standard value of eta that would be used if \code{criterion = "eta"} (e.g. 0.4),
       but much smaller value of epsilon than would be used if \code{criterion = "epsilon"}, e.g. 0.02.}
}}

\item{eta}{A positive numeric value for the minimal mutual distance of
changes, relative to bandwidth (if \code{criterion = "eta"} or \code{criterion = "eta.and.epsilon"}).}

\item{epsilon}{a numeric value in (0,1] for the minimal size of exceeding
environments, relative to moving sum bandwidth (if \code{criterion = "epsilon"} or \code{criterion = "eta.and.epsilon"}).}

\item{use.mean}{\code{Logical variable}, only to be used if \code{data.drive.kern.par=TRUE}. If set to \code{TRUE}, the mean
of pairwise distances is used to set the kernel function tuning parameter, instead of the median. May be useful for binary data,
not recommended to be used otherwise.}
}
\value{
A \code{list} object that contains the following fields:
   \item{x}{Input data}
   \item{G}{Moving window bandwidth}
   \item{lag}{Lag used to detect changes}
   \item{kernel.f, data.driven.kern.par, use.mean}{Input parameters}
   \item{kern.par}{The value of the kernel tuning parameter}
   \item{threshold, alpha, reps, boot.dep, boot.method, parallel}{Input parameters}
   \item{threshold.val}{Threshold value for declaring change points}
   \item{criterion, eta, epsilon}{Input parameters}
   \item{test.stat}{A vector containing the NP-MOJO detector statistics computed from the input data}
   \item{cpts}{A vector containing the estimated change point locations}
   \item{p.vals}{The corresponding p values of the change points, if the bootstrap method was used}
}
\description{
For a given lagged value of the time series, performs nonparametric change point detection of a possibly multivariate
time series. If \code{lag} \eqn{\ell = 0}, then only marginal changes are detected.
If \code{lag} \eqn{\ell \neq 0}, then changes in the pairwise distribution of \eqn{(X_t , X_{t+\ell})} are detected.
}
\details{
The single-lag NP-MOJO algorithm for nonparametric change point detection is described in McGonigle, E. T. and Cho, H. (2023)
Nonparametric data segmentation in multivariate time series via joint characteristic functions.  \emph{arXiv preprint arXiv:2305.07581}.
}
\examples{
\donttest{
set.seed(123)
n <- 1000
noise <- c(rep(1, 300), rep(0.4, 700)) * stats::arima.sim(model = list(ar = 0.3), n = 1000)
signal <- c(rep(0, 700), rep(0.5, 300))
x <- signal + noise
x.c <- np.mojo(x, G = 166, lag = 0)
x.c$cpts
x.c$p.vals
}
}
\references{
McGonigle, E.T., Cho, H. (2023). Nonparametric data segmentation in multivariate time series via joint characteristic functions. \emph{arXiv preprint arXiv:2305.07581}.

Fan, Y., de Micheaux, P.L., Penev, S. and Salopek, D. (2017). Multivariate nonparametric test of independence. \emph{Journal of Multivariate Analysis},
153, pp.189-210.
}
\seealso{
\link{np.mojo.multilag}
}
