% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CptNonPar-package.R
\docType{package}
\name{CptNonPar-package}
\alias{CptNonPar}
\alias{CptNonPar-package}
\title{CptNonPar: Nonparametric Change Point Detection for Multivariate Time Series}
\description{
Implements the nonparametric moving sum procedure for detecting changes in the joint characteristic function (NP-MOJO) for multiple change point detection in multivariate time series. See McGonigle, E. T., Cho, H. (2025) \doi{10.1093/biomet/asaf024} for description of the NP-MOJO methodology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EuanMcGonigle/CptNonPar}
  \item Report bugs at \url{https://github.com/EuanMcGonigle/CptNonPar/issues}
}

}
\author{
\strong{Maintainer}: Euan T. McGonigle \email{e.t.mcgonigle@soton.ac.uk}

Authors:
\itemize{
  \item Haeran Cho \email{haeran.cho@bristol.ac.uk}
}

}
\keyword{internal}
